+++ DahuaVTO.py
@@ -1,27 +1,11 @@
 #!/usr/bin/env python3
 
-import os
 import sys
 import logging
 from time import sleep
 
 from clients.DahuaClient import DahuaClient
 from clients.MQTTClient import MQTTClient
-
-DEBUG = str(os.environ.get('DEBUG', False)).lower() == str(True).lower()
-
-log_level = logging.DEBUG if DEBUG else logging.INFO
-
-root = logging.getLogger()
-root.setLevel(log_level)
-
-stream_handler = logging.StreamHandler(sys.stdout)
-stream_handler.setLevel(log_level)
-formatter = logging.Formatter('%(asctime)s %(levelname)s %(name)s %(message)s')
-stream_handler.setFormatter(formatter)
-root.addHandler(stream_handler)
-
-_LOGGER = logging.getLogger(__name__)
 
 
 class DahuaVTOManager:
@@ -36,6 +20,21 @@ class DahuaVTOManager:
         while True:
             sleep(1)
 
+if __name__ == "__main__":
+    log_level = logging.DEBUG
+    root = logging.getLogger()
+    root.setLevel(log_level)
 
-manager = DahuaVTOManager()
-manager.initialize()
+    stream_handler = logging.StreamHandler(sys.stdout)
+    stream_handler.setLevel(log_level)
+    formatter = logging.Formatter(
+        "'%(asctime)s.%(msecs)04d' " "%(levelname)s %(module)s:%(lineno)d -> %(message)s",
+        "%m/%d/%y %H:%M:%S",
+    )
+    stream_handler.setFormatter(formatter)
+    root.addHandler(stream_handler)
+
+    logger = logging.getLogger(__name__)
+
+    manager = DahuaVTOManager()
+    manager.initialize()